feat: Add Google AdSense banner and update page title with react-helmet
This commit adds a Google AdSense banner to the QR Code Generator React component. It also updates the page title using the react-helmet library. The AdSense banner includes a script tag that is added to the DOM using useEffect, and the client ID and ad slot values are included directly in the component. The page title is set using the Helmet component from react-helmet. This makes the QR Code Generator more dynamic and profitable by displaying relevant ads and improving SEO.